+++ identity-service/src/main/java/co/unicauca/piedrazul/identity/domain/entities/User.java
@@ -17,7 +17,7 @@ public class User {
 
     @Id
     @Column(name = "user_id")
-    private int id;
+    private long id;
 
     @Column(name = "user_username", nullable = false, unique = true, length = 150)
     private String username;
@@ -57,8 +57,8 @@ public class User {
     public User() {}
 
     // --- Getters y Setters ---
-    public int getId() { return id; }
-    public void setId(int id) { this.id = id; }
+    public long getId() { return id; }
+    public void setId(long id) { this.id = id; }
     public String getUsername() { return username; }
     public void setUsername(String username) { this.username = username; }
     public String getPassword() { return password; }
