+++ patient-service/src/main/java/co/unicauca/piedrazul/patient/domain/repository/PatientRepository.java
@@ -14,15 +14,15 @@ import co.unicauca.piedrazul.patient.domain.entities.Patient;
  * @author Santiago Solarte
  */
 @Repository
-public interface PatientRepository extends JpaRepository<Patient, Integer> {
+public interface PatientRepository extends JpaRepository<Patient, Long> {
 
     // Verifica si ya existe un paciente con ese id
-    boolean existsById(int id);
+    boolean existsById(long id);
 
     // Busca pacientes por email (campo pat_email)
     Optional<Patient> findByEmail(String email);
 
     // Busca pacientes por número de documento (id)
     @Query("SELECT p FROM Patient p WHERE p.id = ?1")
-    Optional<Patient> findByDocumentId(int documentId);
+    Optional<Patient> findByDocumentId(long documentId);
 }
